+++ assets/javascripts/kitten/components/atoms/icon-badge/stories.js
@@ -18,13 +18,6 @@ export default {
       </div>
     ),
   ],
-  args: {
-    disabled: false,
-    valid: false,
-    empty: false,
-    size: 'normal',
-    children: <StarIcon color={COLORS.background1} />,
-  },
   argTypes: {
     disabled: {
       name: 'disabled',
@@ -38,6 +31,23 @@ export default {
       name: 'empty',
       control: 'boolean',
     },
+    borderRadius: {
+      name: 'borderRadius',
+      control: 'number',
+    },
+    borderWidth: {
+      name: 'borderWidth',
+      control: 'number',
+    },
+    borderColor: {
+      name: 'borderColor',
+      control: 'color',
+    },
+    borderStyle: {
+      name: 'borderStyle',
+      control: 'select',
+      options: ['solid', 'dotted', 'dashed'],
+    },
     size: {
       name: 'size',
       control: 'select',
@@ -46,4 +56,27 @@ export default {
   },
 }
 
+const args = {
+  disabled: false,
+  valid: false,
+  empty: false,
+  size: 'normal',
+  borderRadius: 30,
+  borderColor: null,
+  borderWidth: null,
+  borderStyle: null,
+  children: <StarIcon color={COLORS.background1} />,
+}
+
 export const Default = args => <IconBadge {...args} />
+
+Default.args = args
+
+export const WithBorder = args => <IconBadge {...args} />
+
+WithBorder.args = {
+  ...args,
+  borderColor: '#222',
+  borderWidth: 2,
+  borderStyle: 'solid',
+}